Make peParser and pePrinter available for other OS.
Also make binDynMod library for binary dynamic modification
functionality.
