+++ cTools/libs/LEB128/CMakeLists.txt
@@ -22,7 +22,7 @@
 
 add_vipon_library(
   NAME LEB128
-  TYPE STATIC
+  TYPE SHARED
   HEADERS LEB128.h
   SOURCES LEB128.c
   INSTALL ON
