+++ cTools/libs/binDynMod/elfDynMod/elf32DynMod.h
@@ -1,7 +1,7 @@
 /***
  * MIT License
  *
- * Copyright (c) 2021 Konychev Valerii
+ * Copyright (c) 2023 Konychev Valerii
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -22,41 +22,20 @@
  * SOFTWARE.
  */
 
-#include "foo.h"
-#include "bar.h"
-#include "test.h"
-#include "elf64Parse.h"
-#include "comdef.h"
+#include "elf32Parse.h"
 
-#ifdef __WIN__
-# undef ERROR
-# define ERROR(...)
-#endif
-
-static void hookFooWithBar(char *argv0)
-{
-    Elf64File *elf64 = elf64Parse(argv0);
-    if (elf64 == NULL) {
-        ERROR("Cannot parse %s", argv0);
-        return;
-    }
-
-    elf64Hook(elf64, "foo", bar);
-    elf64Free(elf64);
-}
-
-int main(int argc, char *argv[])
-{
-    UNUSED(argc);
-
-    foo();
-    bar();
-    hookFooWithBar(argv[0]);
-
-    char *str1 = foo();
-    char *str2 = bar();
-    EXPECT_STR_EQ(str1, str2);
-
-    return 0;
-}
+/***
+ * Before:
+ *  If you need a file position, you should to save it.
+ * Input:
+ *  elf32 - Elf32File structer with initialized field fd.
+ *  func - name of function, that is nedded to hooked.
+ *  hand - address of handler function.
+ * Output:
+ *  Success:
+ *      Old relocation addr.
+ *  Fail:
+ *      NULL point.
+ */
+void *elf32Hook(const Elf32File *elf32, const char *func, const void *hand);
