+++ cTools/libs/containers/CMakeLists.txt
@@ -31,7 +31,7 @@ set(CONTAINERS_SRCS
 
 add_vipon_library(
   NAME containers
-  TYPE STATIC
+  TYPE SHARED
   HEADERS ${CONTAINERS_HEADERS}
   SOURCES ${CONTAINERS_SRCS}
   LINK_LIBS mem
