+++ cTools/libs/string/CMakeLists.txt
@@ -32,7 +32,7 @@ set(STRING_SOURCES
 
 add_vipon_library(
   NAME string
-  TYPE STATIC
+  TYPE SHARED
   HEADERS ${STRING_HEADERS}
   SOURCES ${STRING_SOURCES}
   INSTALL ON
